requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==3.6.4.0
black==23.3.0
celery==5.2.7
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.2.5
cryptography==41.0.2
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==1.3.0
dj-rest-auth==3.0.0
Django==4.2
django-allauth==0.54.0
django-cors-headers==3.14.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
factory-boy==3.2.1
Faker==19.2.0
filelock==3.12.2
flake8==6.0.0
gunicorn==20.1.0
identify==2.5.26
idna==3.4
inflection==0.5.1
isort==5.12.0
kombu==5.3.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.9.1
pre-commit==3.3.2
prompt-toolkit==3.0.39
psycopg==3.1.9
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
ruff==0.0.267
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
virtualenv==20.24.1
wcwidth==0.2.6
whitenoise==6.4.0
